CMake: Enable C in addition to C++
Using the FindThreads with certain versions of CMake requires that C
be enabled for the project as well as C++. This is likely something
that we would want anyway. This should hopefully allow travis to
build (but the unit tests should still fail).

See the issue on the CMake issue tracker here:
    https://cmake.org/Bug/view.php?id=14905