add a Makefile to call the CMake Generated Makefiles
This was done to make building easier and was based Makefile found in the
following stack overflow question.

http://stackoverflow.com/questions/406004/call-cmake-from-make-to-create-makefiles

The answer to the question goes on to suggest a better way to do this sort of
thing, but I couldn't end up getting it working. Also this may not be optimal
became CMake can generate other build files besides make files.