cmake: make fetched libs platform independent
builds where failing on MacOSX because a hard-coded linux shared library
extension of ".so" was expected, but instead the platform default extension
"dylib" was used instead. To fix this the library extensions for both gtest
and anax libraries were replaced with a cmake vaiable that expands to the
platforms suffix.

This also fixes the library prefixes by prefixing libs with another cmake
variable and removes all hard coded "lib" prefixes from libraries were removed
to let cmake handle the platform prefixes