Integrate clang format into CMake
Add a clangformat cmake target that runs clang-format on all of the
project sources. At present we glob for these. It would be possible
to pass them them to the top level CMake file instead, but this is
a simpler solution. The downside is that cmake will need to be run each
time new source files are added to the project.

Also added a .clang-format file with some basic rules. We can play around
with these and see what we want.