CMake: Improve how subdir source files are used
Make the component and system CMake files set a list of source files
rather than defining libraries, which simplifies things.

This also should fix a build failure that occurred due to the systems
library being dependent on the Anax headers, but not declaring a
dependency on Anax. This meant that the Anax git repository would not
necessarily be pulled down before the systems library was built.