Add multiple geese to use texture resource manager
Add multiple geese to main.cpp so that the texture resource manager
is used.

The geese are created by a goose factory function, to prevent needless
duplication of code for each goose.

Also added output messages to the texture resource manager so that
we can see that textures are being re-used correctly.