Fix issue with player movement speed
The PlayerInputSystem was incorrectly basing the player's velocity on
the time since the last update. This resulted in extremely slow
movement speeds.

The correct behavior is that the movement system should take the time
difference into account when calculating position, but that the actual
velocity is independent of time. This may change if we add
inertia/acceleration to movement in the future.