Fix a uninitialized shared_ptr returned by ResourceManagerTexture::get()
Symptoms:
The game was crashing.

Problem:
An uninitialized shared_ptr p_texture was created in the scope of the function
and another p_texture was accidentally created in the inner scope of an
if/else statement causing confusion about what variable would actually be
returned, by the function ResourceManagerTexture::get().

Solution:
Replace the second p_texture with the first so a valid pointer will be
returned.