Render: Fix copying of components
The render system was copying components rather that getting references
to them. This is due to a missing & in the type declaration for the
variables to contain the components. This is not only slower, but would
have resulted in bugs had the system modified these components (as the
actual component would not have been modified).