Data: Add support for camera and player components
Add support for creating the camera and player components from the
JSON data. This means that the factories in world.cpp are no longer
necessary. They are still being left in place for the time being as
they could be useful for developing other features, at least for the
time being.