Add tool to check JSON validity/format
Add a jsoncheck tool that is used to check that the validity
and formatting of the JSON files match what the game itself would
output. Updated existing files to conform to its output style.

The jsoncpp serialization does cause objects to have a trailing
space following the object name. We could update the write method
to strip newlines in the future if desired.